Integration: `2024.w26`

See merge request oai/openairinterface5g!2815

* !2750 Add Ubuntu 24.04 to/retire Ubuntu 18.04 from supported distributions, use Ubuntu 22 for CI container images
* !2793 Option '-O' in phy simulators to provide a configuration file
* !2797 Increase TX_DATA.request TB buffer size, fix compute_PDU_Length
* !2814 Update documentation and FEATURE SET
* !2720 Remove thread-pool Makefile
* remove extra log

+++ nfapi/open-nFAPI/nfapi/public_inc/nfapi_nr_interface_scf.h
@@ -26,10 +26,11 @@
 // Extension to the generic structures for single tlv values
 
 typedef struct {
-  /// Value: 0 -> 1, 0: Payload is carried directly in the value field, 1: Pointer to payload is in the value field 
-  uint16_t tag; 
-  /// Length of the actual payload in bytes, without the padding bytes Value: 0 → 65535
-  uint16_t length;
+  /// Value: 0 -> 1, 0: Payload is carried directly in the value field, 1: Pointer to payload is in the value field
+  uint16_t tag;
+  /// Length of the actual payload in bytes, without the padding bytes Value: 0 → 0xFFFFFFFF
+  /// Note: This change to uint32_t is a deviation from version 10.02 of the SCF222 standard
+  uint32_t length;
   union { 
     uint32_t *ptr;
     uint32_t direct[38016];
